Add files via upload
Using the Java classes DatagramPacket and DatagramSocket to develop the UDP mode communication.
A UDP  server will accept and handle any client.
The server needs to be a concurrent server; therefore multiple clients are handled while operating.
Finally, a disconnect message must be sent to the clients when the server is about to shutdown i.e. a 5 second warning.
Any client still connected will receive a force disconnect and their connection to the server is released.